+++ src/cheat/precompiled.h
@@ -280,6 +280,7 @@ namespace hl
 #include "features/movement/GroundStrafe.h"
 #include "features/movement/EdgeBug.h"
 #include "features/movement/StrafeHack.h"
+#include "features/movement/StrafeHelper.h"
 #include "features/movement/Movement.h"
 
 // ui